+++ go/string_commands_test.go
@@ -4,6 +4,7 @@ package glide
 
 import (
 	"context"
+	"encoding/json"
 	"fmt"
 	"time"
 
@@ -610,7 +611,7 @@ func ExampleClient_LCS() {
 	if err != nil {
 		fmt.Println("Glide example failed with an error: ", err)
 	}
-	fmt.Println(result)
+	fmt.Println(result.MatchString)
 
 	// LCS is only available in 7.0 and above. It will fail in any server < 7.0
 
@@ -625,7 +626,7 @@ func ExampleClusterClient_LCS() {
 	if err != nil {
 		fmt.Println("Glide example failed with an error: ", err)
 	}
-	fmt.Println(result)
+	fmt.Println(result.MatchString)
 
 	// LCS is only available in 7.0 and above. It will fail in any release < 7.0
 
@@ -676,7 +677,7 @@ func ExampleClient_LCSLen() {
 	if err != nil {
 		fmt.Println("Glide example failed with an error: ", err)
 	}
-	fmt.Println(result)
+	fmt.Println(result.Len)
 
 	// LCS is only available in 7.0 and above. It will fail in any server < 7.0
 
@@ -693,7 +694,7 @@ func ExampleClusterClient_LCSLen() {
 	if err != nil {
 		fmt.Println("Glide example failed with an error: ", err)
 	}
-	fmt.Println(result)
+	fmt.Println(result.Len)
 
 	// LCS is only available in 7.0 and above. It will fail in any release < 7.0
 
@@ -712,7 +713,8 @@ func ExampleClient_LCSWithOptions() {
 	if err != nil {
 		fmt.Println("Glide example failed with an error: ", err)
 	}
-	fmt.Println("Basic LCS result:", result1)
+	jsonRes1, _ := json.Marshal(result1)
+	fmt.Println("Basic LCS result:", string(jsonRes1))
 
 	// LCS IDX with MINMATCHLEN = 4
 	optsWithMin := options.NewLCSIdxOptions()
@@ -721,34 +723,44 @@ func ExampleClient_LCSWithOptions() {
 	if err != nil {
 		fmt.Println("Glide example failed with an error: ", err)
 	}
-	fmt.Println("With MinMatchLen 4:", result2)
+	jsonRes2, _ := json.Marshal(result2)
+	fmt.Println("With MinMatchLen 4:", string(jsonRes2))
 
 	// LCS is only available in 7.0 and above. It will fail in any server < 7.0
 
 	// Output:
-	// Basic LCS result: map[len:3 matches:[[0 1 0 1] [6 7 4 5]]]
-	// With MinMatchLen 4: map[len:0 matches:[]]
+	// Basic LCS result: {"MatchString":"","Matches":[{"Key1":{"Start":4,"End":7},"Key2":{"Start":5,"End":8},"MatchLen":4},{"Key1":{"Start":2,"End":3},"Key2":{"Start":0,"End":1},"MatchLen":2}],"Len":6}
+	// With MinMatchLen 4: {"MatchString":"","Matches":[{"Key1":{"Start":4,"End":7},"Key2":{"Start":5,"End":8},"MatchLen":4}],"Len":6}
 }
 
 func ExampleClusterClient_LCSWithOptions() {
 	var client *ClusterClient = getExampleClusterClient() // example helper function
 
-	client.Set(context.Background(), "{my_key}1", "ohmytext")
-	client.Set(context.Background(), "{my_key}2", "mynewtext")
+	client.Set(context.Background(), "{key}1", "ohmytext")
+	client.Set(context.Background(), "{key}2", "mynewtext")
 
-	// LCS IDX with MINMATCHLEN and WITHMATCHLEN
+	// Basic LCS IDX without additional options
 	opts := options.NewLCSIdxOptions()
-	opts.SetMinMatchLen(2)
-	opts.SetWithMatchLen(true)
-	result, err := client.LCSWithOptions(context.Background(), "{my_key}1", "{my_key}2", *opts)
+	result1, err := client.LCSWithOptions(context.Background(), "{key}1", "{key}2", *opts)
 	if err != nil {
 		fmt.Println("Glide example failed with an error: ", err)
 	}
+	jsonRes1, _ := json.Marshal(result1)
+	fmt.Println("Basic LCS result:", string(jsonRes1))
 
-	fmt.Println("Full result with both options:", result)
+	// LCS IDX with MINMATCHLEN = 4
+	optsWithMin := options.NewLCSIdxOptions()
+	optsWithMin.SetMinMatchLen(4)
+	result2, err := client.LCSWithOptions(context.Background(), "{key}1", "{key}2", *optsWithMin)
+	if err != nil {
+		fmt.Println("Glide example failed with an error: ", err)
+	}
+	jsonRes2, _ := json.Marshal(result2)
+	fmt.Println("With MinMatchLen 4:", string(jsonRes2))
 
-	// LCS is only available in 7.0 and above. It will fail in any release < 7.0
+	// LCS is only available in 7.0 and above. It will fail in any server < 7.0
 
 	// Output:
-	// Full result with both options: map[len:3 matches:[[0 1 0 1 2] [6 7 4 5 2]]]
+	// Basic LCS result: {"MatchString":"","Matches":[{"Key1":{"Start":4,"End":7},"Key2":{"Start":5,"End":8},"MatchLen":4},{"Key1":{"Start":2,"End":3},"Key2":{"Start":0,"End":1},"MatchLen":2}],"Len":6}
+	// With MinMatchLen 4: {"MatchString":"","Matches":[{"Key1":{"Start":4,"End":7},"Key2":{"Start":5,"End":8},"MatchLen":4}],"Len":6}
 }