+++ CHANGELOG.md
@@ -127,6 +127,7 @@
 * Go: Change `ZAddIncr` Return Type to be `float64` ([#4190](https://github.com/valkey-io/valkey-glide/pull/4190))
 * Go: Update return type for `LMPop` and related commands ([#4166](https://github.com/valkey-io/valkey-glide/pull/4166))
 * Go: `ZRankWithScore` and `ZRevRankWithScore` update response type ([#4196](https://github.com/valkey-io/valkey-glide/pull/4196))
+* Go: Update return type for LCS commands ([#4187](https://github.com/valkey-io/valkey-glide/pull/4187))
 
 #### Fixes